requirements.txt
Flask==2.0.1
requests==2.26.0
python-dotenv==0.19.1
gspread==5.0.0
oauth2client==4.1.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.12.2
gunicorn==20.1.0 